use uint64_t for seeding

+++ oc.cpp
@@ -3,7 +3,7 @@
 constexpr char OC::possible_chars[];
 
 std::string OC::random_sequence(int length) {
-		unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
+		uint64_t seed = std::chrono::system_clock::now().time_since_epoch().count();
 		std::mt19937 generator(seed);
 
 		std::string result;
